Extract gift redemption state code into separate classes
