+++ support-workers/src/main/scala/com/gu/support/workers/lambdas/CreateZuoraSubscription.scala
@@ -13,6 +13,7 @@ import com.gu.support.promotions.{PromoError, PromotionService}
 import com.gu.support.redemption.corporate.GetCodeStatus.RedemptionInvalid
 import com.gu.support.redemption._
 import com.gu.support.redemption.corporate.{DynamoLookup, DynamoUpdate, GetCodeStatus, RedemptionTable, SetCodeStatus}
+import com.gu.support.redemption.gifting.{GiftRedemptionState, RedeemedInThisRequest, SubscriptionRedemptionState, Unredeemed}
 import com.gu.support.redemption.gifting.generator.GiftCodeGeneratorService
 import com.gu.support.redemptions.RedemptionData
 import com.gu.support.workers._
@@ -255,27 +256,6 @@ object CreateZuoraSubscription {
 
 object DigitalSubscriptionGiftRedemption {
 
-  val expirationTimeInMonths = 12
-
-  sealed abstract class SubscriptionRedemptionState(val clientCode: String)
-
-  object Unredeemed {
-    val clientCode = "unredeemed"
-  }
-
-  case class Unredeemed(subscriptionId: String) extends SubscriptionRedemptionState(Unredeemed.clientCode)
-
-  // This can happen if Zuora is responding very slowly - a redemption request may succeed but not return a response
-  // until after the CreateZuoraSubscription lambda has timed out meaning that the redemption will be retried with the
-  // same requestId. In this case we want the lambda to succeed so that we progress to the next lambda
-  case object RedeemedInThisRequest extends SubscriptionRedemptionState("redeemed_in_this_request")
-
-  case object Redeemed extends SubscriptionRedemptionState("redeemed")
-
-  case object Expired extends SubscriptionRedemptionState("expired")
-
-  case object NotFound extends SubscriptionRedemptionState("not_found")
-
   def maybeDigitalSubscriptionGiftRedemption(product: ProductType, paymentMethod: Either[PaymentMethod, RedemptionData]) = {
     product match {
       case d: DigitalPack if paymentMethod.isRight && d.readerType == Gift => paymentMethod.right.toOption
@@ -292,26 +272,12 @@ object DigitalSubscriptionGiftRedemption {
   ): Future[HandlerResult[SendThankYouEmailState]] =
     zuoraService.getSubscriptionFromRedemptionCode(redemptionData.redemptionCode).flatMap(
       redemptionQueryResponse =>
-        getSubscriptionState(redemptionQueryResponse, state.requestId.toString) match {
+       GiftRedemptionState.getSubscriptionState(redemptionQueryResponse, state.requestId.toString) match {
           case Unredeemed(subscriptionId) => redeemInZuora(subscriptionId, state, redemptionData, requestInfo, zuoraService, catalogService)
           case RedeemedInThisRequest => Future.fromTry(buildHandlerResult(UpdateRedemptionDataResponse(true), state, redemptionData, requestInfo))
           case otherState: SubscriptionRedemptionState => Future.failed(new RuntimeException(otherState.clientCode))
         }
     )
-
-  def getSubscriptionState(existingSub: SubscriptionRedemptionQueryResponse, requestId: String) =
-    existingSub.records match {
-      case existingSubFields :: Nil if existingSubFields.contractEffectiveDate.plusMonths(expirationTimeInMonths).isBefore(LocalDate.now()) =>
-        Expired
-      case existingSubFields :: Nil if existingSubFields.gifteeIdentityId.isEmpty =>
-        Unredeemed(existingSubFields.id)
-      case existingSubFields :: Nil if existingSubFields.createdRequestId == requestId =>
-        RedeemedInThisRequest
-      case _ :: Nil =>
-        Redeemed
-      case _ =>
-        NotFound
-    }
 
 
   private def redeemInZuora(
