+++ support-workers/src/main/scala/com/gu/zuora/subscriptionBuilders/DigitalSubscriptionBuilder.scala
@@ -11,6 +11,7 @@ import com.gu.support.config.{TouchPointEnvironment, ZuoraDigitalPackConfig}
 import com.gu.support.promotions.{PromoCode, PromoError, PromotionService}
 import com.gu.support.redemption.corporate.GetCodeStatus.{InvalidReaderType, RedemptionInvalid}
 import com.gu.support.redemption.corporate.GetCodeStatus
+import com.gu.support.redemption.gifting.GiftRedemptionState
 import com.gu.support.redemption.gifting.generator.CodeBuilder.GiftCode
 import com.gu.support.redemption.gifting.generator.GiftCodeGeneratorService
 import com.gu.support.redemptions.{RedemptionCode, RedemptionData}
@@ -71,7 +72,7 @@ object DigitalSubscriptionBuilder {
   )(implicit ec: ExecutionContext): BuildResult = {
 
     val (contractAcceptanceDelay, autoRenew, initialTerm) = if (readerType == Gift)
-      (0, false, DigitalSubscriptionGiftRedemption.expirationTimeInMonths + 1)
+      (0, false, GiftRedemptionState.expirationTimeInMonths + 1)
     else
       (purchase.config.defaultFreeTrialPeriod + purchase.config.paymentGracePeriod, true, 12)
