+++ support-workers/src/test/scala/com/gu/support/workers/integration/DigitalSubscriptionGiftRedemptionIntegrationSpec.scala
@@ -4,11 +4,11 @@ import java.util.UUID
 
 import com.gu.salesforce.Fixtures.idId
 import com.gu.support.redemption.corporate.GetCodeStatus.{CodeAlreadyUsed, NoSuchCode}
+import com.gu.support.redemption.gifting.{NotFound, Redeemed}
 import com.gu.support.redemption.gifting.generator.CodeBuilder.GiftCode
 import com.gu.support.redemption.gifting.generator.GiftCodeGeneratorService
 import com.gu.support.redemptions.{RedemptionCode, RedemptionData}
 import com.gu.support.workers.JsonFixtures.{createDigiPackGiftRedemptionJson, createDigiPackGiftSubscriptionJson}
-import com.gu.support.workers.lambdas.DigitalSubscriptionGiftRedemption.{NotFound, Redeemed}
 import com.gu.support.workers.{Annual, AsyncLambdaSpec, BillingPeriod, Fixtures, MockContext, RequestInfo}
 import com.gu.support.workers.lambdas.{DigitalSubscriptionGiftRedemption, HandlerResult}
 import com.gu.support.workers.states.{CreateZuoraSubscriptionState, SendThankYouEmailState}
